Allow expressions to generate instrs before them
This is used for the "function call" expression, as it needs to call
the function first, save the result to a temporary value and then put
a reference to that value as the expression.

The test added in this commit tests that a function can call another
function as an expression and return it's result.