Implelemt int+int instrinsic
This also introduces a depdency on the trivia package for pattern
matching.

The intrinsics work by looking up an intrinsic dispatcher from the
function name, and letting that dispatcher return the runtime
definition of the intrinsic function, which is a runtime-function with
it's body set to a labmda. When evaluate-instr sees such a call, it
just calls that lambda instead of generating a call instruction.